requirements.txt
pytest
pytest-docker
pytest-xdist
allure-pytest
selenium
webdriver-manager
requests
sqlalchemy
pydantic